+++ CHANGELOG.md
@@ -1,20 +1,64 @@
- - 0.0.7
+# ChangeLog
 
-- Update unit tests
-- Update README.md
-- Pretty printer
+- 0.0.36 - 0.0.37
+- update tests
+- update commands to correctly use package manager
+- change templates
+- update linter
+- update docs
 
-- 0.0.8
-- Update dev commands
+-0.0.34 - 0.0.35
+- update docker compose command
+- update docs
+- auth wip
 
-- 0.0.9
-- Updates to vanty dev commands
 
-- 0.0.10
-- Update README.md
+- 0.0.30 - 32
+- update docs
+
+- 0.0.29
+- update docs
+
+- 0.0.28
+- update error with admin
+
+- 0.0.27
+- update docs
+- update running tests
+- add admin for show configs
+- add community for creating new apps
+
+- 0.0.26
+- update tests
+
+- 0.0.18 - 0.0.25
+- Add template in env.
+- Template will scaffold new app including api, tests and types.
+- fix bugs in scaffold
+- add url option
+
+- prev >> 0.0.17
+- Version checking
+- Update template .env
+- Update migrate
+- fix broken optional str
+
+- 0.0.12
+- Fix regression in vanty dev commands
 
 - 0.0.11
 - update dev commands
 
-- 0.0.12
-- Fix regression in vanty dev commands
+- 0.0.10
+- Update README.md
+
+- 0.0.9
+- Updates to vanty dev commands
+
+- 0.0.8
+- Update dev commands
+
+- 0.0.7
+- Update unit tests
+- Update README.md
+- Pretty printer